fix: Combine MemberControllerTests (#23)
